+++ l10n_br_sale/models/__init__.py
@@ -3,5 +3,5 @@
 
 from . import res_company
 from . import res_config_settings
-from . import sale
-from . import sale_line
+from . import sale_order
+from . import sale_order_line
