+++ l10n_br_sale/models/sale_order_line.py
@@ -10,6 +10,11 @@ class SaleOrderLine(models.Model):
     _name = 'sale.order.line'
     _inherit = [_name, 'l10n_br_fiscal.document.line.mixin']
 
+    def _get_protected_fields(self):
+        protected_fields = super(SaleOrderLine, self)._get_protected_fields()
+        return protected_fields + [
+            'fiscal_tax_ids', 'fiscal_operation_id', 'fiscal_operation_line_id']
+
     @api.model
     def _default_fiscal_operation(self):
         return self.env.user.company_id.sale_fiscal_operation_id
@@ -30,37 +35,31 @@ class SaleOrderLine(models.Model):
         relation='fiscal_sale_line_tax_rel',
         column1='document_id',
         column2='fiscal_tax_id',
-        string='Fiscal Taxes',
-    )
+        string='Fiscal Taxes')
 
     quantity = fields.Float(
         related='product_uom_qty',
-        depends=['product_uom_qty'],
-    )
+        depends=['product_uom_qty'])
 
     uom_id = fields.Many2one(
         related='product_uom',
-        depends=['product_uom'],
-    )
+        depends=['product_uom'])
 
     tax_framework = fields.Selection(
         selection=TAX_FRAMEWORK,
-        related='order_id.company_id.tax_framework',
-        string='Tax Framework',
-    )
+        related="order_id.company_id.tax_framework",
+        string="Tax Framework")
 
     partner_id = fields.Many2one(
         comodel_name='res.partner',
         related='order_id.partner_id',
-        string='Partner',
-    )
+        string='Partner')
 
     # Add Fields in model sale.order.line
     price_gross = fields.Monetary(
         compute='_compute_amount',
         string='Gross Amount',
-        default=0.00,
-    )
+        default=0.00)
 
     comment_ids = fields.Many2many(
         comodel_name='l10n_br_fiscal.comment',
@@ -69,13 +68,6 @@ class SaleOrderLine(models.Model):
         column2='comment_id',
         string='Comments',
     )
-
-    def _get_protected_fields(self):
-        protected_fields = super()._get_protected_fields()
-        return protected_fields + [
-            'fiscal_tax_ids', 'fiscal_operation_id',
-            'fiscal_operation_line_id',
-        ]
 
     @api.depends(
         'product_uom_qty',
@@ -90,24 +82,25 @@ class SaleOrderLine(models.Model):
         'tax_id')
     def _compute_amount(self):
         """Compute the amounts of the SO line."""
-        super()._compute_amount()
-        for l in self:
-            l._update_taxes()
-            price_tax = l.price_tax + l.amount_tax_not_included
-            price_total = (
-                l.price_subtotal + l.freight_value +
-                l.insurance_value + l.other_costs_value)
+        super(SaleOrderLine, self)._compute_amount()
+        for line in self:
+            line._update_taxes()
+            price_tax = line.price_tax + line.amount_tax_not_included
+            price_subtotal = (
+                line.price_subtotal + line.freight_value +
+                line.insurance_value + line.other_costs_value)
 
-            l.update({
+            line.update({
                 'price_tax': price_tax,
-                'price_gross': l.price_subtotal + l.discount_value,
-                'price_total': price_total + price_tax,
+                'price_gross': price_subtotal + line.discount_value,
+                'price_subtotal': price_subtotal,
+                'price_total': price_subtotal + price_tax,
             })
 
     @api.multi
     def _prepare_invoice_line(self, qty):
         self.ensure_one()
-        result = super()._prepare_invoice_line(qty)
+        result = super(SaleOrderLine, self)._prepare_invoice_line(qty)
         result.update(self._prepare_br_fiscal_dict())
         return result
 
@@ -118,24 +111,21 @@ class SaleOrderLine(models.Model):
         self._onchange_commercial_quantity()
 
     @api.onchange('discount')
-    def _onchange_discount_percent(self):
+    def _onchange_discount(self):
         """Update discount value"""
-        if not self.env.user.has_group('l10n_br_sale.group_discount_per_value'):
+        if self.env.user.has_group('l10n_br_sale.group_discount_per_value'):
             if self.discount:
                 self.discount_value = (
-                    (self.product_uom_qty * self.price_unit)
-                    * (self.discount / 100)
-                )
+                    self.price_gross * (self.discount / 100))
 
     @api.onchange('discount_value')
     def _onchange_discount_value(self):
         """Update discount percent"""
-        if self.env.user.has_group('l10n_br_sale.group_discount_per_value'):
+        if self.env.user.has_group('l10n_br_sale.group_total_discount'):
             if self.discount_value:
                 self.discount = ((self.discount_value * 100) /
                                  (self.product_uom_qty * self.price_unit))
 
     @api.onchange('fiscal_tax_ids')
     def _onchange_fiscal_tax_ids(self):
-        super()._onchange_fiscal_tax_ids()
         self.tax_id |= self.fiscal_tax_ids.account_taxes()
